converted over to dev_t, uid_t, gid_t

+++ include/fcntl.h
@@ -6,7 +6,7 @@
 #include <fs.h>
 #include <sys/fcntl.h>
 
-struct flock *create_flock(int type, int whence, int start, int len);
+struct flock *create_flock(int type, int whence, off_t start, size_t len);
 int can_flock(struct inode *file, struct flock *l);
 int disengage_flock(struct inode *file, struct flock *l);
 int engage_flock(struct inode *file, struct flock *l, int);
